% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_chisq.R
\name{.as_count_matrix}
\alias{.as_count_matrix}
\title{Coerce a count input to an integer-valued matrix.}
\usage{
.as_count_matrix(counts, n_rows, n_cols)
}
\description{
Coerce a count input to an integer-valued matrix.
}
\keyword{internal}
