% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_chisq.R
\name{.chisq_one}
\alias{.chisq_one}
\title{Dispatch a single test method on a single table.}
\usage{
.chisq_one(tab, method, fisher_simulate, fisher_B)
}
\description{
Dispatch a single test method on a single table.
}
\keyword{internal}
