% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_chisq.R
\name{.hyper_point}
\alias{.hyper_point}
\title{Point (hypergeometric) probability of an observed 2x2 table.}
\usage{
.hyper_point(tab)
}
\description{
Point (hypergeometric) probability of an observed 2x2 table.
}
\keyword{internal}
