% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_way_between_anova.R
\name{.round_output_rw}
\alias{.round_output_rw}
\title{Internal: rounding wrapper using {roundwork}}
\usage{
.round_output_rw(x, digits, method)
}
\arguments{
\item{x}{numeric}

\item{digits}{integer}

\item{method}{one of "half_up", "half_down", "bankers", "truncate"}
}
\description{
Internal: rounding wrapper using {roundwork}
}
\keyword{internal}
