% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_independent_t.R
\name{plot_multiverse_d}
\alias{plot_multiverse_d}
\title{Plot multiverse of Cohen's d / Hedges' g and 95\% CIs}
\usage{
plot_multiverse_d(res)
}
\arguments{
\item{res}{A list returned by \code{\link{recalc_independent_t_d}} or
\code{\link{recalc_independent_t}}.}
}
\value{
A ggplot object.
}
\description{
Plot multiverse of Cohen's d / Hedges' g and 95\% CIs
}
