% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_independent_t.R
\name{plot_multiverse_p}
\alias{plot_multiverse_p}
\title{Plot multiverse of p-values}
\usage{
plot_multiverse_p(res)
}
\arguments{
\item{res}{A list returned by \code{\link{recalc_independent_t_p}} or
\code{\link{recalc_independent_t}}.}
}
\value{
A ggplot object.
}
\description{
Plot multiverse of p-values
}
