% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{recalc}
\alias{recalc-package}
\alias{recalc}
\title{recalc}
\description{
Recalculate statistical results from reported summary and test statistics.
}
\section{Important}{

Package is in early stage development.
}

\author{
\strong{Maintainer}: Ian Hussey \email{Ian.hussey@unibe.ch} (\href{https://orcid.org/0000-0001-8906-7559}{ORCID})

Authors:
\itemize{
  \item Wallrich Lukas \email{lukas.wallrich@gmail.com} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID})
}

}
