% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{recalc_result}
\alias{recalc_result}
\title{Standard return shape: tibble with reported and recomputed intervals}
\usage{
recalc_result(check, reported_value, reported_int, recomputed_int)
}
\description{
Standard return shape: tibble with reported and recomputed intervals
}
\keyword{internal}
