% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_t_test_summary.R
\name{plot_multiverse_p}
\alias{plot_multiverse_p}
\title{Plot multiverse of p-values from independent_t_test_summary}
\usage{
plot_multiverse_p(res)
}
\arguments{
\item{res}{A list returned by independent_t_test_summary(),
containing an element \code{p_results} with column \code{p_unrounded}.}
}
\value{
A ggplot object visualizing the distribution of recalculated
unrounded p-values across all analytic choices.
}
\description{
Plot multiverse of p-values from independent_t_test_summary
}
\examples{
\dontrun{
plot_multiverse_p(res)
}

}
