% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc-package.R
\docType{package}
\name{recalc-package}
\alias{recalc}
\alias{recalc-package}
\title{recalc: Tools for Recalculation Multiverses}
\description{
This package provides functions to explore multiverses of effect sizes,
confidence intervals, and p-values compatible with reported summary
statistics.
}
\author{
\strong{Maintainer}: Ian Hussey \email{Ian.hussey@unibe.ch} (\href{https://orcid.org/0000-0001-8906-7559}{ORCID})

Authors:
\itemize{
  \item Wallrich Lukas \email{lukas.wallrich@gmail.com} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID})
}

}
\keyword{internal}
